backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic==2.9.0
pydantic-settings==2.5.2
alembic==1.13.3
python-dotenv==1.0.1
anthropic==0.39.0
httpx==0.24.1
tenacity==9.0.0
